+++ telemetry/collector/src/main/java/ru/practicum/event/mapstruct/HubEventMapper.java
@@ -1,4 +1,4 @@
-package ru.practicum.event.mupstruct;
+package ru.practicum.event.mapstruct;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
