+++ telemetry/collector/src/main/resources/application.properties
@@ -1,2 +1,2 @@
 kafka.bootstrap-servers=localhost:9092
-kafka.value-serializer=ru.practicum.event.producer.AvroSerializer
+kafka.value-serializer=event_serializer.AvroSerializer
