+++ telemetry/serialization/avro-schemas/src/main/java/event_serializer/AvroSerializer.java
@@ -1,4 +1,4 @@
-package ru.practicum.event.producer;
+package event_serializer;
 
 import org.apache.avro.io.BinaryEncoder;
 import org.apache.avro.io.DatumWriter;
